Column factory uses 64 bit length values

+++ lib/fsttable/columnfactory.h
@@ -11,37 +11,37 @@
 class ColumnFactory : public IColumnFactory
 {
 	// Inherited via IColumnFactory
-	IFactorColumn* CreateFactorColumn(int nrOfRows, int nrOfLevels, FstColumnAttribute columnAttribute)
+	IFactorColumn* CreateFactorColumn(uint64_t nrOfRows, int nrOfLevels, FstColumnAttribute columnAttribute)
 	{
 		return new FactorVectorAdapter(nrOfRows, nrOfLevels, columnAttribute);
 	}
 
-	ILogicalColumn* CreateLogicalColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+	ILogicalColumn* CreateLogicalColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
 	{
 		return new LogicalVectorAdapter(nrOfRows);
 	}
 
-	IInt64Column* CreateInt64Column(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+	IInt64Column* CreateInt64Column(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
 	{
 		return new Int64VectorAdapter(nrOfRows, columnAttribute, scale);
 	}
 
-	IDoubleColumn* CreateDoubleColumn(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+	IDoubleColumn* CreateDoubleColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
 	{
 		return new DoubleVectorAdapter(nrOfRows, columnAttribute, scale);
 	}
 
-	IIntegerColumn* CreateIntegerColumn(int nrOfRows, FstColumnAttribute columnAttribute, short int scale)
+	IIntegerColumn* CreateIntegerColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute, short int scale)
 	{
 		return new IntVectorAdapter(nrOfRows, columnAttribute, scale);
 	}
 
-	IByteColumn* CreateByteColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+	IByteColumn* CreateByteColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
 	{
 		return new ByteVectorAdapter(nrOfRows, columnAttribute);
 	}
 
-	IStringColumn* CreateStringColumn(int nrOfRows, FstColumnAttribute columnAttribute)
+	IStringColumn* CreateStringColumn(uint64_t nrOfRows, FstColumnAttribute columnAttribute)
 	{
 		return new StringColumn();
 	}
